[fix]관리자 마일리지 조회 페이지 API 연동
